+++ .metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.rse.core.prefs
@@ -1,3 +1,4 @@
 activeuserprofiles=Sparky;Team
 eclipse.preferences.version=1
 org.eclipse.rse.systemtype.local.systemType.defaultUserId=Jon
+useridperkey=Sparky.Local\=Jon;
