+++ ChatServerThread.java
@@ -1,55 +1,75 @@
+
 import java.net.*;
 import java.io.*;
-public class ChatServerThread extends Thread{
-	   private ChatServer       server    = null;
-	   private Socket           socket    = null;
-	   private int              ID        = -1;
-	   private DataInputStream  streamIn  =  null;
-	   private DataOutputStream streamOut = null;
-	   
 
-	   public ChatServerThread(ChatServer _server, Socket _socket)
-	   {  super();
-	      server = _server;
-	      socket = _socket;
-	      ID     = socket.getPort();
+public class ChatServerThread extends Thread{
+	private ChatServer server = null;
+	private Socket socket = null;
+	private int ID = -1;
+	private DataInputStream streamIn = null;
+	private DataOutputStream streamOut = null;
+	private  boolean running = true;
+	public volatile int opponentID = -1;
+	public volatile String username = null;
+	public volatile User user = null;
+   
+
+   public ChatServerThread(ChatServer _server, Socket _socket){
+	   super();
+	   server = _server;
+	   socket = _socket;
+	   ID = socket.getPort();
+   }
+   
+   public void send(String msg){
+	   try{
+		   streamOut.writeUTF(msg);
+		   streamOut.flush();
+       }
+       catch(IOException ioe){
+    	   System.out.println(ID + " ERROR sending: " + ioe.getMessage());
+    	   server.remove(ID);
+    	   interrupt();
+       }
+   }
+   
+   public int getID(){
+	   return ID;
+   }
+   
+   public void run(){
+	   System.out.println("Server Thread " + ID + " running.");
+	   send("\n-----------------------------");
+	   send("Welcome to the King Me Server");
+	   send("-----------------------------");
+	   send("\nPlease enter your name:");
+	   while (running){
+		   try{
+			   server.handle(ID, streamIn.readUTF());
+		   }
+		   catch(IOException ioe){
+			   System.out.println(ID + " ERROR reading: " + ioe.getMessage());
+			   server.remove(ID);
+			   interrupt();
+		   }
+		   catch(ArrayIndexOutOfBoundsException e){ continue; }
 	   }
-	   public void send(String msg)
-	   {   try
-	       {  streamOut.writeUTF(msg);
-	          streamOut.flush();
-	       }
-	       catch(IOException ioe)
-	       {  System.out.println(ID + " ERROR sending: " + ioe.getMessage());
-	          server.remove(ID);
-	          stop();
-	       }
-	   }
-	   public int getID()
-	   {  return ID;
-	   }
-	   public void run()
-	   {  System.out.println("Server Thread " + ID + " running.");
-	      while (true)
-	      {  try
-	         {  server.handle(ID, streamIn.readUTF());
-	         }
-	         catch(IOException ioe)
-	         {  System.out.println(ID + " ERROR reading: " + ioe.getMessage());
-	            server.remove(ID);
-	            stop();
-	         }
-	      }
-	   }
-	   public void open() throws IOException
-	   {  streamIn = new DataInputStream(new 
-	                        BufferedInputStream(socket.getInputStream()));
-	      streamOut = new DataOutputStream(new
-	                        BufferedOutputStream(socket.getOutputStream()));
-	   }
-	   public void close() throws IOException
-	   {  if (socket != null)    socket.close();
-	      if (streamIn != null)  streamIn.close();
-	      if (streamOut != null) streamOut.close();
-	   }
+   }
+   
+   public void open() throws IOException{
+	   streamIn = new DataInputStream(new 
+                        BufferedInputStream(socket.getInputStream()));
+	   streamOut = new DataOutputStream(new
+                        BufferedOutputStream(socket.getOutputStream()));
+   }
+   
+   public void close() throws IOException{
+	   if (socket != null)
+		   socket.close();
+	   if (streamIn != null)
+		   streamIn.close();
+	   if (streamOut != null)
+		   streamOut.close();
+	   running = false;
+   }
 }
